+++ Makefile
@@ -1,22 +1,14 @@
-.PHONY: build, test, coverage, clean, mocks
+.PHONY: build, test, mocks
 
 build:
 	@echo "Building..."
-	@go build -o dist/etherfi ./cmd/etherfi/
+	@go build -o avs-cli ./bin/avs-cli/
 
 test:
 	@echo "Testing..."
 	@go test -v ./...
 
-coverage:
-	@echo "Testing with coverage..."
-	@go test -coverprofile=coverage.out ./...
-	#@go tool cover -html=coverage.out
-
 mocks:
 	@echo "Generating mocks..."
 	@mockery --recursive --output ./mocks --name BLSAvsSigner
 
-clean:
-	@echo "Cleaning..."
-	@rm -rf dist
