+++ bin/avs-cli/README.md
