+++ bin/avs-cli/eigenda/eigenda.go
