+++ bin/avs-cli/eigenda/prepare-registration.go
