+++ bin/avs-cli/eigenda/register.go
