+++ bin/avs-cli/eoracle/eoracle.go
