+++ bin/avs-cli/eoracle/eoracle_test.go
