+++ bin/avs-cli/eoracle/prepareRegistration.go
