+++ bin/avs-cli/eoracle/register.go
