+++ bin/avs-cli/helpers.go
