+++ bin/avs-cli/main.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/dsrvlabs/etherfi-avs-operator-tool/cmd/etherfi/eigenda"
-	"github.com/dsrvlabs/etherfi-avs-operator-tool/cmd/etherfi/eoracle"
-	"github.com/dsrvlabs/etherfi-avs-operator-tool/cmd/etherfi/witness-chain"
+	"github.com/dsrvlabs/etherfi-avs-operator-tool/bin/avs-cli/eigenda"
+	"github.com/dsrvlabs/etherfi-avs-operator-tool/bin/avs-cli/eoracle"
+	"github.com/dsrvlabs/etherfi-avs-operator-tool/bin/avs-cli/witness-chain"
 	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/urfave/cli/v3"
 )
