+++ bin/avs-cli/operator_details.go
