+++ bin/avs-cli/operator_details_test.go
