+++ bin/avs-cli/registerBLS.go
