+++ bin/avs-cli/registerOperator.go
