+++ bin/avs-cli/registerOperator_test.go
