+++ bin/avs-cli/registrationDigest.go
