+++ bin/avs-cli/registrationDigest_test.go
