+++ bin/avs-cli/registration_utils.go
