+++ bin/avs-cli/signing.go
