+++ bin/avs-cli/update_ecdsa_signer.go
