+++ bin/avs-cli/update_whitelist.go
