+++ bin/avs-cli/witness-chain/prepareRegistration.go
