+++ bin/avs-cli/witness-chain/register.go
