+++ bin/avs-cli/witness-chain/witness.go
